+++ src/cloudtypes.jl
@@ -3,60 +3,64 @@
 """
 # struct CloudMetadata{T<:AbstractFloat} <: CloudTrack{T}
 
-Metadata struct for cloud track data with fields `ID`, `date` (date range `start` to `stop`),
-`area`, `flex`, `useLON`, and `file` similar to `FlightMetadata`.
+Metadata struct for cloud track data with fields `id`, `date` (date range `start` to `stop`),
+`area`, `flex`, `use_lon`, `root` and `file`.
+
+See also: [`FlightMetadata`](@ref), [`CloudData`](@ref), [`CloudTrack`](@ref), and [`CloudSet`](@ref)
 
 ## Instantiation
 
 Use modified constructor to calculate `area` and `date` range from `time` and
 `lat`/`lon` vectors in `data`.
 
-    function CloudMetadata{T}(
-        ID::Union{Int,AbstractString},
+    CloudMetadata{T}(
         data::DataFrame,
+        id::Int32,
         flex::Tuple{Vararg{NamedTuple{(:range, :min, :max),Tuple{UnitRange{Int},T,T}}}},
-        useLON::Bool,
-        file::AbstractString
-    ) where T
+        use_lon::Bool,
+        root::UInt16,
+        file::UInt16
+    ) where T<:AbstractFloat
 
 Or hand over individual fields to the unmodified constructor.
 
-    function CloudMetadata{T}(
-        ID::String,
+    CloudMetadata{T}(
+        id::Int32,
         date::NamedTuple{(:start,:stop),Tuple{DateTime,DateTime}},
         area::NamedTuple{(:latmin,:latmax,:elonmin,:elonmax,:wlonmin,:wlonmax),NTuple{6,T}},
         flex::Tuple{Vararg{NamedTuple{(:range, :min, :max),Tuple{UnitRange{Int},T,T}}}},
-        useLON::Bool,
-        file::String
+        use_lon::Bool,
+        root::UInt16,
+        file::UInt16
     ) where T
 """
 struct CloudMetadata{T<:AbstractFloat} <: CloudTrack{T}
-    ID::String
+    id::Int32
     date::NamedTuple{(:start,:stop),Tuple{DateTime,DateTime}}
     area::NamedTuple{(:latmin,:latmax,:elonmin,:elonmax,:wlonmin,:wlonmax),NTuple{6,T}}
     flex::Tuple{Vararg{NamedTuple{(:range, :min, :max),Tuple{UnitRange{Int},T,T}}}}
-    useLON::Bool
-    file::String
+    use_lon::Bool
+    root::UInt16
+    file::UInt16
 end
 
 #* Main constructor for CloudMetadata with auto
 function CloudMetadata{T}(
-    ID::Union{Int,AbstractString},
-    time::Vector{DateTime},
-    lat::Vector{T},
-    lon::Vector{T},
+    data::DataFrame,
+    id::Int32,
     flex::Tuple{Vararg{NamedTuple{(:range, :min, :max),Tuple{UnitRange{Int},T,T}}}},
-    useLON::Bool,
-    file::AbstractString
+    use_lon::Bool,
+    root::UInt16,
+    file::UInt16
 ) where T<:AbstractFloat
     # T = promote_type(eltype(data.lat), eltype(data.lon))
-    elonmax = isempty(lon[lon.≥0]) ? T(NaN) : maximum(lon[lon.≥0])
-    elonmin = isempty(lon[lon.≥0]) ? T(NaN) : minimum(lon[lon.≥0])
-    wlonmax = isempty(lon[lon.<0]) ? T(NaN) : maximum(lon[lon.<0])
-    wlonmin = isempty(lon[lon.<0]) ? T(NaN) : minimum(lon[lon.<0])
-    area = (latmin=minimum(lat), latmax=maximum(lat),
+    elonmax = isempty(data.lon[data.lon.≥0]) ? T(NaN) : T(maximum(data.lon[data.lon.≥0]))
+    elonmin = isempty(data.lon[data.lon.≥0]) ? T(NaN) : T(minimum(data.lon[data.lon.≥0]))
+    wlonmax = isempty(data.lon[data.lon.<0]) ? T(NaN) : T(maximum(data.lon[data.lon.<0]))
+    wlonmin = isempty(data.lon[data.lon.<0]) ? T(NaN) : T(minimum(data.lon[data.lon.<0]))
+    area = (latmin=T(minimum(data.lat)), latmax=T(maximum(data.lat)),
     elonmin=elonmin, elonmax=elonmax, wlonmin=wlonmin, wlonmax=wlonmax)
-    CloudMetadata{T}(ID, (start=time[1], stop=time[end]), area, flex, useLON, file)
+    CloudMetadata{T}(id, (start=data.time[1], stop=data.time[end]), area, flex, use_lon, root, file)
 end #constructor 2 CloudMetadata
 
 #* Default constructor for type promotion to Float32
@@ -66,16 +70,16 @@ CloudMetadata(args...) = CloudMetadata{Float32}(args...)
 CloudMetadata{T}() where T<:AbstractFloat = CloudMetadata{T}(
     "", (start=Dates.now(), stop=Dates.now()), (latmin=T(NaN), latmax=T(NaN),
     elonmin=T(NaN), elonmax=T(NaN), wlonmin=T(NaN), wlonmax=T(NaN)),
-    ((range=0:0, min=T(NaN), max=T(NaN)),), false, ""
+    ((range=0:0, min=T(NaN), max=T(NaN)),), false, 0x0000, 0x0000
 )
 
 #* Constructor for type promotion from CloudMetadata with different float precision
 CloudMetadata{T}(meta::CloudMetadata) where T<:AbstractFloat = CloudMetadata{T}(
-    meta.ID, meta.date, (latmin = T(meta.area.latmin), latmax = T(meta.area.latmax),
+    meta.id, meta.date, (latmin = T(meta.area.latmin), latmax = T(meta.area.latmax),
     elonmin = T(meta.area.elonmin), elonmax = T(meta.area.elonmax),
     wlonmin = T(meta.area.wlonmin), wlonmax = T(meta.area.wlonmax)),
     Tuple([(range = m.range, min = T.(m.min), max = T.(m.max)) for m in meta.flex]),
-    meta.useLON, meta.file
+    meta.use_lon, meta.root, meta.file
 )
 
 
@@ -86,15 +90,24 @@ CloudMetadata{T}(meta::CloudMetadata) where T<:AbstractFloat = CloudMetadata{T}(
 
 Store Data related to a single cloud track.
 
+See also: [`CloudMetadata`](@ref), [`CloudTrack`](@ref), [`CloudSet`](@ref), and [`FlightData`](@ref)
+
 ## Fields
 
 - `time::Vector{DateTime}`
 - `lat::Vector{T}`
 - `lon::Vector{T}`
-- `metadata::CloudMetadata`
+- `metadata::CloudMetadata{T}`
 
 Default floating point precision is `Float32`. There are basic validity checks
 during instantiation.
+
+## Instantiation
+
+The main constructor constructs `CloudData` from a `DataFrame` with columns `time`, `lat`, and
+`lon`:
+
+    CloudData{T}(data::DataFrame, metadata::CloudMetadata) where T
 """
 struct CloudData{T<:AbstractFloat} <: PrimaryTrack{T}
     time::Vector{DateTime}
@@ -103,7 +116,9 @@ struct CloudData{T<:AbstractFloat} <: PrimaryTrack{T}
     metadata::CloudMetadata{T}
 
     #* Unmodified constructor for `CloudData` with basic checks for correct `data``
-    function CloudData{T}(data::DataFrame, metadata::CloudMetadata{T}) where T<:AbstractFloat
+    function CloudData{T}(
+        time::Vector{DateTime}, lat::Vector{T}, lon::Vector{T}, metadata::CloudMetadata{T}
+        ) where T<:AbstractFloat
         # Column checks and warnings
         length(time) == length(lat) == length(lon) ||
             throw(DimensionMismatch("all data vectors must have the same length in CloudData"))
@@ -113,6 +128,14 @@ struct CloudData{T<:AbstractFloat} <: PrimaryTrack{T}
         new{T}(time, lat, lon, metadata)
     end #constructor 1 CloudTrack
 end #struct CloudTrack
+
+#* Main constructor for `CloudData`
+function CloudData{T}(
+    data::DataFrame,
+    metadata::CloudMetadata
+)::CloudData{T} where T<:AbstractFloat
+    CloudData{T}(data.time, data.lat, data.lon, CloudMetadata{T}(metadata))
+end #constructor 2 CloudData
 
 #* Default constructor for type promotion to Float32
 CloudData(args...) = CloudData{Float32}(args...)
@@ -141,8 +164,9 @@ CloudTrack(args...) = CloudData{Float32}(args...)
 # struct CloudSet{T<:AbstractFloat} <: PrimarySet{T}
 
 Database for cloud track data with fields:
-- `tracks::Vector{CloudTrack}`
-- `metadata::PrimaryMetadata`
+
+- `tracks::StructArray{CloudData{T}}`
+- `metadata::PrimaryMetadata{T}`
 
 # Instantiation
 
@@ -153,10 +177,10 @@ and optional remarks.
 
 Or use the unmodified constructor.
 
-    CloudSet{T}(tracks::Vector{CloudData{T}}, metadata::PrimaryMetadata{T}) where T
+    CloudSet{T}(tracks::StructArray{CloudData{T}}, metadata::PrimaryMetadata{T}) where T
 """
 struct CloudSet{T<:AbstractFloat} <: PrimarySet{T}
-    tracks::Vector{CloudData{T}}
+    tracks::StructArray{CloudData{T}}
     metadata::PrimaryMetadata{T}
 end
 
@@ -171,11 +195,14 @@ function CloudSet{T}(
     # Track computing time
     tstart = Dates.now()
     # Scan folders for mat files
-    roots = ds.OrderedDict{String,UInt16}()
-    paths = findfiles!(roots, collect(folders), ".mat")
+    pathdict = ds.OrderedDict{String,ds.OrderedDict}(
+        "roots" => ds.OrderedDict{String,UInt16}(),
+        "files" => ds.OrderedDict{String,UInt16}()
+    )
+    paths = findfiles!(pathdict, collect(folders), ".mat")
 
-    # Load cloud tracks from mat files into TrackMatcher in Julia format
-    tracks = load_cloudtracks(paths; structname, Float=T)
+    #* Load cloud tracks from mat files into TrackMatcher in Julia format
+    tracks = load_cloudtracks(paths, pathdict, structname, T)
     # Calculate load time
     tend = Dates.now()
     tc = tz.ZonedDateTime(tend, tz.localzone())
@@ -184,58 +211,41 @@ function CloudSet{T}(
     # For now find min/max times in CloudTracks
     tmin = minimum(t.time[1] for t in tracks)
     tmax = maximum(t.time[end] for t in tracks)
+    # Define lookup dictionary for metadata
+    roots = ds.OrderedDict{UInt16,String}(v => k for (k, v) in pathdict["roots"])
+    files = ds.OrderedDict{UInt16,String}(v => k for (k, v) in pathdict["files"])
+    pathdict = ds.OrderedDict{String,ds.OrderedDict}("roots" => roots, "files" => files)
 
     @info string("CloudSet loaded in ",
         "$(join(loadtime.periods[1:min(2,length(loadtime.periods))], ", ")) to",
         "\n▪ tracks ($(length(tracks)) entries)\n▪ metadata")
 
     # Instantiate CloudSet
-    CloudSet{T}(tracks, PrimaryMetadata{T}(NaN, (start=tmin, stop=tmax), tc, loadtime, attachments))
+    CloudSet{T}(tracks, PrimaryMetadata{T}(0, (start=tmin, stop=tmax), ds.OrderedDict(), pathdict, tc, loadtime, attachments))
 end #modified constructor 2
 
-"""
-    CloudSet{T}() where T
-
-External constructor for empty `CloudSet{T}`.
-"""
+#* Constructor for empty CloudSet
 CloudSet{T}() where T<:AbstractFloat = CloudSet{T}(CloudData{T}[], PrimaryMetadata{T}())
 
-"""
-    CloudSet(args...; kwargs...)
-
-Default `CloudSet` constructor for single floating point precision.
-"""
+#* Default constructor for type promotion to Float32
 CloudSet(args...; kwargs...) = CloudSet{Float32}(args...; kwargs...)
 
-"""
-    CloudSet{T}(cloud::CloudSet) where T
-
-External `CloudSet` constructor for floating point conversions.
-"""
+#* Constructor for type promotion from CloudSet with different float precision
 CloudSet{T}(cloud::CloudSet) where T<:AbstractFloat = CloudSet{T}(
     CloudData{T}.(cloud.tracks),
     PrimaryMetadata{T}(cloud.metadata)
 )
 
 """
-    PrimarySet{T}(tracks::Vector{CloudData{T}}, metadata::PrimaryMetadata{T}) where T
+    PrimarySet{T}(tracks::StructArray{CloudData{T}}, metadata::PrimaryMetadata{T}) where T<:AbstractFloat
+    PrimarySet{T}(folders::String...; kwargs...) where T<:AbstractFloat
 
-Alias constructor for CloudSet{T} unmodified constructor.
+Alias constructor for `CloudSet{T}` constructors.
 """
-PrimarySet{T}(tracks::Vector{CloudData{T}}, metadata::PrimaryMetadata{T}) where T<:AbstractFloat =
+function PrimarySet end
+
+PrimarySet{T}(tracks::StructArray{CloudData{T}}, metadata::PrimaryMetadata{T}) where T<:AbstractFloat =
     CloudSet{T}(tracks, metadata)
 
-"""
-    PrimarySet{T}(folders::String...; remarks=nothing) where T
-
-Alias constructor for `CloudSet{T}`.
-"""
-PrimarySet{T}(folders::String...; remarks=nothing) where T<:AbstractFloat =
-    CloudSet{T}(folders; remarks)
-
-"""
-    PrimarySet{T}(tracks::CloudSet) where T
-
-Alias constructor for `CloudSet` with converted floating point precision.
-"""
-PrimarySet{T}(tracks::CloudSet) where T<:AbstractFloat = CloudSet{T}(tracks)
+PrimarySet{T}(folders::String...; kwargs...) where T<:AbstractFloat =
+    CloudSet{T}(folders...; kwargs...)